% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbn.network.diagram.R
\name{bbn.network.diagram}
\alias{bbn.network.diagram}
\title{Bayesian Belief Model v4.0.1 - Network diagrams}
\usage{
bbn.network.diagram(
  bbn.network,
  font.size = 0.7,
  arrow.size = 4,
  arrange = layout_on_sphere
)
}
\arguments{
\item{bbn.network.diagram}{To illustrate a ‘complex system’.}
}
\value{
It visualises all nodes and interactions in a full network.
}
\description{
27 Feb 2024
Copyright Rick Stafford

Requires all functions loaded and:
  1) BBN interaction grid - named bbn.model
  2) list of scenarios indicating changes (1-12 can be used). Named priors1, priors2, ... priors12
Nodes can also be colour coded by theme.
It is easiest to create a slightly different csv file to produce these networks, which allows for the colour coding.

Required arguments:
bbn.network - a csv file as described above, with note paid to the first three column names.

Optional arguments:
   1) font.size - default = 0.7. Changes the font in the figure produced.
      The value here is a multiplier of the default font size used in the igraph package and does not correspond to the font.size argument in the bbn.timeseries() function.
   2) arrow.size - default = 4. Changes the size of the arrows.
      Note, sizes do vary based on interaction strength, so this is a multiplier for visualisation purposes.
      Negative interactions are shown by red arrows, and positive interactions by black arrows.
   3) arrange - this describes how the final diagram looks.
      Default is `layout_on_sphere` but `layout_on_grid` provides the same layout as in the bbn.visualise() function and ensures nodes are structured in the order specified in the network.
      Other layouts, including `layout_on_sphere` are more randomly determined, and better/clearer diagrams may occur if you run these multiple times.
      Other options are from the igraph package
}
