% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbn.sensitivity.R
\name{bbn.sensitivity}
\alias{bbn.sensitivity}
\title{Bayesian Belief Model v4.0.1 - Sensitivity function}
\usage{
bbn.sensitivity(bbn.model, boot_max = 1000, ...)
}
\arguments{
\item{bbn.sensitivity}{}
}
\value{
This function produces a list of the most important parameters/interaction strengths to examine.
}
\description{
27 Feb 2024
Copyright Rick Stafford

Requires all functions loaded and:
  1) BBN interaction grid - named bbn.model
  2) list of scenarios indicating changes (1-12 can be used). Named priors1, priors2, ... priors12

The function works by bootstrapping with multiple changes to prior values and interaction strengths in the network.
The frequency shows the number of times a modified interaction shows up as important in causing a change to the listed nodes.
As such, those interactions showing as more frequent in the table or figure are likely to be most influential in any predictions.
These should be subject to closer scrutiny in terms of values used.
(note, this does not mean the values are incorrect or should be reduced from more extreme values - i.e. from 4 to 3, just that they should be carefully checked)

Required arguments:
  1) bbn.model - a matrix or dataframe of interactions between different model nodes.
     One or more nodes (recommended no more than 3) which would be the main outcomes of interest in the model.
     The spelling of these nodes needs to be identical (including capital letters) to that in the imported csv file (note, you should include spaces if these are in your csv file, rather than the dot notation used once imported into R).

Optional arguments:
  2) boot_max - the number of bootstraps to perform.
     Suggested range for exploratory analysis 100-1000.
     For final analysis recommended size = 1000 - 10000 - note, this can take a long time to run. Default value is 1000.
}
