% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.R
\docType{data}
\name{combined}
\alias{combined}
\title{Combined Treatment Data}
\format{
A data frame with 9 rows and 2 columns:
\describe{
  \item{Increase}{integer}
  \item{Node}{Variable names}
}
}
\source{
<https://doi.org/10.1016/j.ocecoaman.2015.04.013>
}
\description{
This dataset represents the numerical changes in species populations on a rocky shore
ecosystem due to the combined treatment of removing dogwhelks and adding periwinkles.
It reflects the complex interactions and potential synergistic effects
of multiple ecological interventions.
}
