% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogwhelk.R
\docType{data}
\name{dogwhelk}
\alias{dogwhelk}
\title{Dogwhelk Removal Data}
\format{
A data frame with 9 rows and 2 columns:
\describe{
  \item{Increase}{integer}
  \item{Node}{Variable names}
}
}
\source{
<https://doi.org/10.1016/j.ocecoaman.2015.04.013>
}
\description{
This dataset represents the numerical changes in species populations on
a rocky shore ecosystem due to the removal of dogwhelks.
It provides insights into the potential ecological impacts of removing
a predatory species.
}
