% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsEmpty.R
\name{isEmpty}
\alias{isEmpty}
\title{Check if an Object is Empty}
\usage{
isEmpty(x)
}
\arguments{
\item{x}{The object to check for emptiness.}
}
\value{
A logical value: \code{TRUE} if the object is empty, \code{FALSE} otherwise.
}
\description{
This function determines whether the provided object is empty.
}
\details{
\code{isEmpty()} checks if the given object, \code{x}, has a length of 0,
indicating that it is empty. It can be used with various types of objects
in R, including vectors, lists, and data frames.
}
\examples{
# Check an empty vector
isEmpty(c())

# Check a non-empty vector
isEmpty(c(1, 2, 3))

# Check an empty list
isEmpty(list())

# Check a non-empty list
isEmpty(list(a = 1, b = 2))

# Check an empty data frame
isEmpty(data.frame())

# Check a non-empty data frame
isEmpty(mtcars)

}
