% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPANetwork.R
\docType{data}
\name{MPANetwork}
\alias{MPANetwork}
\title{Simple model of MPA, ecological components and management}
\format{
A data frame with 11 rows and 12 columns:
\describe{
  \item{id}{Variable names}
  \item{Lobster.fishery}{integer}
  \item{Finfish.fishery}{integer}
  \item{Fish.density}{integer}
  \item{Seals}{integer}
  \item{Lobster.Recruitment }{integer}
  \item{Divers}{integer}
  \item{Spiny.lobster}{integer}
  \item{Lobster}{integer}
  \item{Snails}{integer}
  \item{Algae}{integer}
  \item{Revenue}{integer}
}
}
\source{
<unpublished work by Rick Stafford>
}
\description{
This dataset represents an interaction model of marine protected area
and ecological components
This is an example dataset loosely based on Lundy Island MCZ.
}
