% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoPotting.R
\docType{data}
\name{NoPotting}
\alias{NoPotting}
\title{Dataset represents banning potting (for crabs / lobsters etc)
in a Marine Protected Area
Data presents insights into how management measures may affect ecological communities
in MPAs}
\format{
A data frame with 11 rows and 2 columns:
\describe{
  \item{Increase}{integer}
  \item{Node}{Variable names}
}
}
\source{
<unpublished work by Rick Stafford>
}
\description{
Dataset represents banning potting (for crabs / lobsters etc)
in a Marine Protected Area
Data presents insights into how management measures may affect ecological communities
in MPAs
}
