% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoTake_data.R
\docType{data}
\name{NoTake_data}
\alias{NoTake_data}
\title{Dataset represents banning all fishing in a Marine Protected Area}
\format{
A data frame with 11 rows and 2 columns:
\describe{
  \item{Increase}{integer}
  \item{Node}{Variable names}
}
}
\source{
<unpublished work by Rick Stafford>
}
\description{
Data presents insights into how management measures may affect
ecological communities in MPAs
}
