% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbn.timeseries.R
\name{bbn.timeseries}
\alias{bbn.timeseries}
\title{Bayesian Belief Model v4.0.1 - Time series}
\usage{
bbn.timeseries(bbn.model, priors1, timesteps = 5, disturbance = 1)
}
\arguments{
\item{bbn.timeseries}{}
}
\value{
Graph of each node in the network, visualised across the different timesteps in the model.
}
\description{
27 Feb 2024
Copyright Rick Stafford

Requires all functions loaded and:
  1) BBN interaction grid - named bbn.model
  2) list of scenarios indicating changes (1-12 can be used). Named priors1, priors2, ... priors12

Required arguments:
   1) bbn.model - a matrix or dataframe of interactions between different model nodes.
   2) priors1 - an X by 2 array of initial changes to the system under investigation.
      The first column should be a -4 to 4 (including 0) integer value for each node in the network with negative values indicating a decrease and positive values representing an increase.
      0 represents no change.

Optional Arguments:
   3) timesteps - default = 5. This is the number of timesteps the model performs.
      Note, timesteps are arbitrary and non-linear. However, something occurring in timestep 2, should occur before timestep 3.
   4) disturbance - default = 1.
      where 1 - creates a prolonged or press disturbance as per the bbn.predict() function. Essentially prior values for each manipulated node are at least maintained (if not increased through reinforcement in the model) over all timesteps.
      where 2 - shows a brief pulse disturbance, which can be useful to visualise changes as peaks and troughs in increase and decrease of nodes can propagate through the network
}
